Conflicts:
	Projects/Win32Framework/Win32Framework.vcxproj.filters

Former-commit-id: 895e01cbfafd4bdafc5e8d57d5d9c7fe7d64a448

+++ Sources/Internal/FileSystem/YamlNode.cpp
@@ -123,8 +123,8 @@ uint32 YamlNode::GetCount() const
 {
     switch (GetType())
     {
-        case TYPE_MAP: return objectMap->unordered.size();
-        case TYPE_ARRAY: return objectArray->array.size();
+        case TYPE_MAP: return static_cast<uint32>(objectMap->unordered.size());
+        case TYPE_ARRAY: return static_cast<uint32>(objectArray->array.size());
         default: break;
     }
     return 0;//string nodes does not contain content
@@ -368,7 +368,7 @@ VariantType YamlNode::AsVariantType() const
         if(innerTypeName == DAVA::VariantType::TYPENAME_BYTE_ARRAY)
         {
             const Vector<YamlNode*> &byteArrayNoodes = it->second->AsVector();
-            int32 size = byteArrayNoodes.size();
+            int32 size = static_cast<int32>(byteArrayNoodes.size());
             uint8* innerArray = new uint8[size];
             for (int32 i = 0; i < size; ++i)
             {
