+++ Sources/Internal/Scene3D/MeshInstanceNode.cpp
@@ -670,7 +670,7 @@ void MeshInstanceNode::AddLightmap(int32 polygonGroupIndex, const LightmapData &
 
 void MeshInstanceNode::AddLightmap(const LightmapData & lightmapData)
 {
-	AddLightmap(lightmaps.size(), lightmapData);
+	AddLightmap(static_cast<int32>(lightmaps.size()), lightmapData);
 }
 
 void MeshInstanceNode::ClearLightmaps()
@@ -705,7 +705,7 @@ MeshInstanceNode::LightmapData * MeshInstanceNode::GetLightmapDataForIndex(int32
 
 int32 MeshInstanceNode::GetLightmapCount()
 {
-    return lightmaps.size();
+    return static_cast<int32>(lightmaps.size());
 }
     
 void MeshInstanceNode::CreateDynamicShadowNode()
